fix json double
