+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-# 5.4.5 (2020-10-09)
+# 5.4.5 (2020-10-11)
 
 ### 新特性
 * 【core   】     ConsoleTable代码优化（pr#190@Gitee）
@@ -18,6 +18,7 @@
 ### Bug修复
 * 【core   】     解决农历判断节日未判断大小月导致的问题（issue#I1XHSF@Gitee）
 * 【core   】     解决ListUtil计算总量可能的int溢出问题（pr#1150@Github）
+* 【json   】     解决JSON中转换为double小数精度丢失问题（pr#192@Gitee）
 
 -------------------------------------------------------------------------------------------------------------
