+++ hutool-json/src/main/java/cn/hutool/json/InternalJSONUtil.java
@@ -140,12 +140,14 @@ final class InternalJSONUtil {
 	 * @return A simple JSON value.
 	 */
 	protected static Object stringToValue(String string) {
+		// null处理
 		if (null == string || "null".equalsIgnoreCase(string)) {
 			return JSONNull.NULL;
 		}
 
-		if (StrUtil.EMPTY.equals(string)) {
-			return string;
+		// boolean处理
+		if (0 == string.length()) {
+			return StrUtil.EMPTY;
 		}
 		if ("true".equalsIgnoreCase(string)) {
 			return Boolean.TRUE;
@@ -154,17 +156,22 @@ final class InternalJSONUtil {
 			return Boolean.FALSE;
 		}
 
-		/* If it might be a number, try converting it. If a number cannot be produced, then the value will just be a string. */
+		// Number处理
 		char b = string.charAt(0);
 		if ((b >= '0' && b <= '9') || b == '-') {
 			try {
-				if (string.indexOf('.') > -1 || string.indexOf('e') > -1 || string.indexOf('E') > -1) {
+				if (StrUtil.containsAnyIgnoreCase(string, ".", "e")) {
+					// pr#192@Gitee，Double会出现小数精度丢失问题，此处使用BigDecimal
+					//double d = Double.parseDouble(string);
+					//if (false == Double.isInfinite(d) && false == Double.isNaN(d)) {
+					//	return d;
+					//}
 					return new BigDecimal(string);
 				} else {
-					Long myLong = new Long(string);
-					if (string.equals(myLong.toString())) {
-						if (myLong == myLong.intValue()) {
-							return myLong.intValue();
+					final long myLong = Long.parseLong(string);
+					if (string.equals(Long.toString(myLong))) {
+						if (myLong == (int) myLong) {
+							return (int) myLong;
 						} else {
 							return myLong;
 						}
@@ -173,6 +180,8 @@ final class InternalJSONUtil {
 			} catch (Exception ignore) {
 			}
 		}
+
+		// 其它情况返回原String值下
 		return string;
 	}
